exercicio 48 reescrito codigo - treino
